Devise installation
First, the gem is added to the project's Gemfile:

gem 'devise'

The command is run in the terminal:

bundle install

Run a devise setup builder in the terminal:

rails generate devise:install

or

rails g devise:install

In this way you will have all the necessary configurations for the devise to work.

Follow the necessary recommendations that the command indicates.

===============================================================================

Depending on your application's configuration some manual setup may be required:

  1. Ensure you have defined default url options in your environments files. Here
     is an example of default_url_options appropriate for a development environment
     in config/environments/development.rb:

       config.action_mailer.default_url_options = { host: 'localhost', port: 3000 }

     In production, :host should be set to the actual host of your application.

     * Required for all applications. *

  2. Ensure you have defined root_url to *something* in your config/routes.rb.
     For example:

       root to: "home#index"

     * Not required for API-only Applications *

  3. Ensure you have flash messages in app/views/layouts/application.html.erb.
     For example:

       <p class="notice"><%= notice %></p>
       <p class="alert"><%= alert %></p>

     * Not required for API-only Applications *

  4. You can copy Devise views (for customization) to your app by running:

       rails g devise:views

     * Not required *

===============================================================================

Flash messages added in application view. Which will allow us to send the user messages in case of correct accesses or errors.

Generate a devise model, with the command:

rails generate devise User

or

rails g devise User

Which will allow us to authenticate users.

Execute the migration of the database generated by the User model.

rails db:migrate