Using the Devise gem
Add the before_action element to the Main controller. Devise offers a helper called :authenticate_user!, which verifies that there is a session, if there is no session it will redirect to the login.

Restart the server and reload the localhost to view the changes.

Register a user and verify that they are logged in.

It can be checked with the command in the terminal:

rails c

Which will allow us to view all registered users with the command:

User.all

To exit the debugger, use the command:

exit;

With this, a basic authentication with devise has been created in Ruby on Rails.